Refactor nxos_facts to support resource module (#58705)
